[#22408] CDCSDK: Handle race condition in dynamic table creation.
Summary:
There is a possible race condition when during dynamic table creation, the tablet creation is not completed i.e the retention barriers have not yet been set, but the background thread proceeds and adds the table to the stream. In such a case till the time the tablet creation succeeds, the retention barriers are not set on the tablets of this table, even though it has been added to the stream.

This was observed as flakiness in the test `TestRetentionBarrierRaceWithUpdatePeersAndMetrics` .

This diff introduces a fix for this race condition by calling GetTableLocations on the table before adding it to the stream. If the tablets are not yet initialized then this will fail and the addition of the table to the stream will take place in the next round of the background task.
Jira: DB-11311

Test Plan: Jenkins: test regex: .*CDCSDKConsumptionConsistentChangesTest.*

Reviewers: asrinivasan, stiwary, skumar

Reviewed By: asrinivasan

Subscribers: ybase, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35160