[#22474] YSQL: Remove costs from yb_guc regression test
Summary:
Costs were added to show whether or not disable_cost was added, but costs can be finicky and
change (especially on upgrade to PG15).

All we need to know for each of these test cases is the scan type used, and whether or not the
scan type was disabled. These can be returned by a function to avoid any cost output needing to be
fixed in the test.
Jira: DB-11390

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressMisc#testPgRegressMiscSerial'
```

Reviewers: jason

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D35237