[PLAT-13886] Adds support for running node exporter as user level systemd
Summary:
Adds support for running node exporter as user level systemd.
We will start running node_exporter as user level systemd - so that we can install this during the configure phase.
This will help simplify the onprem manual deployment as post this change we won't need the sudo user for installing the same.

Corresponding change in ansible prometheus are as part of - https://github.com/yugabyte/ansible-prometheus/pull/12

Test Plan:
Created alma based universe.
Verified that the node_exporter is running as user level service.
Started the workload on the universe.
Verified that the metrics are working as expected.

Created centos based universe.
Verified that the node_exporter is running as system level service.
Started the workload on the universe.
Verified that the metrics are working as expected.

Reviewers: amalyshev, anijhawan, sanketh

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35128