[#22520] YSQL: fix logic regarding syscatalog tables
Summary:
Some areas of the code intend to focus on tables that are stored in the
system catalog, but there is no function to represent such tables.
These areas do a crude attempt at that, but it is not consistent and
often has holes.

Add new functions YbIsSysCatalogTabletRelation and
YbIsSysCatalogTabletRelationByIds to consolidate the logic to one place.

There is not much difference that the user will notice.  The following
commands will now return new warnings "unexpected user relation in
system namespace":

- CREATE TABLE information_schema.badtable (i int);
- CREATE INDEX ON information_schema.badtable (i);

The following command will now fail (with "HASH columns must be present
to split by number of tablets"):

- CREATE TABLE information_schema.badtable2 (i int) SPLIT INTO 5 TABLETS;

Preventing these commands entirely is out of scope and noted in issue
#22481.  Once that issue is fixed, these commands should all fail, so
don't bother adding test coverage for now.
Jira: DB-11447

Test Plan:
jenkins

Backport-through: 2.14
Close: #22520

Reviewers: myang, kfranz

Reviewed By: myang

Subscribers: mtakahara, yql

Differential Revision: https://phorge.dev.yugabyte.com/D35271